feat: implementing crud method to UserService

+++ backend/src/main/java/com/moutfit/dto/product/ProductRequestDTO.java
@@ -2,5 +2,5 @@ package com.moutfit.dto.product;
 
 import java.math.BigDecimal;
 
-public record ProductRequestDTO(Long id, String name, String description, BigDecimal price) {
+public record ProductRequestDTO(Integer id, String name, String description, BigDecimal price) {
 }
