+++ backend/src/main/java/com/moutfit/dto/product/ProductResponseDTO.java
@@ -2,5 +2,5 @@ package com.moutfit.dto.product;
 
 import java.math.BigDecimal;
 
-public record ProductResponseDTO(Long id,String name, String description, BigDecimal price) {
+public record ProductResponseDTO(Integer id,String name, String description, BigDecimal price) {
 }
