+++ backend/src/main/java/com/moutfit/dto/product/ProductUpdate.java
@@ -2,5 +2,5 @@ package com.moutfit.dto.product;
 
 import java.math.BigDecimal;
 
-public record ProductUpdate(Long id, String name, String description, BigDecimal price) {
+public record ProductUpdate(Integer id, String name, String description, BigDecimal price) {
 }
